Simplify codebase by removing unnecessary packages
