--- docs/README.md
@@ -1,5 +1,0 @@
-## Image
-
-- Header `README` image by [Sam McGhee](https://unsplash.com/@sammcghee), taken from
-  [Unsplash](https://unsplash.com/)
-- https://unsplash.com/photos/KieCLNzKoBo
