--- docs/pull_request_template.md
@@ -1,11 +1,0 @@
-### Problem
-
-...
-
-### Solution
-
-...
-
-### Notes
-
-- None
