+++ go.mod
@@ -3,6 +3,5 @@ module github.com/revett/youtube-comments
 go 1.12
 
 require (
-	github.com/revett/go-yt-comments v0.0.0-20170726172518-bc385d5d5fc3
 	github.com/stretchr/testify v1.4.0
 )
