--- internal/constants.go
@@ -1,7 +1,0 @@
-package internal
-
-const (
-	// MaxResults is the maximum number of items returned in a single request.
-	MaxResults      = "100"
-	defaultEndpoint = "https://www.googleapis.com/youtube/v3/commentThreads"
-)
